feat: update code

+++ Assets/GameEngine/PvPData.cs
@@ -39,7 +39,7 @@ public class PvPData : Singleton<PvPData>
         OpenedAdvantage = packet.GetIntArray("oppened_advantage");
         SelectingAdvantage = packet.GetInt("selecting_advantage");
         Rank = packet.GetInt("rank");
-        Elo = packet.GetInt("elo");
+        Elo = packet.GetLong("elo");
 
         DefenseCrew.NewFromSFSObject(packet);
     }
